:recycle: splite SettingsEditor to components
